Enum problem_probability::CommandLineMode

pub(crate) enum CommandLineMode {
    Plot {
        bins: usize,
        output: Option<String>,
    },
    Probability {
        only_statistics: bool,
        output: Option<String>,
    },
    Cost {
        all_strategies: bool,
        only_statistics: bool,
        optimizer_fraction: usize,
        output: Option<String>,
    },
    DepGroups,
    Tree,
}

Variants

Plot

Plot a previous measurement

Fields of Plot

bins: usize

Number of bins for displaying the histograms

output: Option<String>

Output where to place the resulting HTML file

Probability

Approximate the probability of failing the migration scenario

Fields of Probability

only_statistics: bool

Store only statistics in the output file.

output: Option<String>

Output where to place the measurement results.

Cost

Compare the cost (in terms of traffic shifts) for the migration scenario.

Fields of Cost

all_strategies: bool

Compare all strategies. If this is set to false, we only use the tree optimizer

only_statistics: bool

Store only statistics in the output file.

optimizer_fraction: usize

Factor to divide the number of iterations by, when computing the cost for the optimizers.

output: Option<String>

Output where to place the measurement results.

DepGroups

Check how the dependnecy gorups strategy performs

Tree

Check how the Tree strategy performs